requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Face recognition and computer vision
opencv-python>=4.8.0
Pillow>=10.0.0
insightface>=0.7.3

# Data handling
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0

# Configuration
pyyaml>=6.0
tqdm>=4.65.0

# Utilities
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking
